csrf protection to citation delete related #609

+++ src/Ojs/JournalBundle/Controller/CitationController.php
@@ -12,6 +12,7 @@ use Ojs\JournalBundle\Form\CitationType;
 use APY\DataGridBundle\Grid\Column\ActionsColumn;
 use APY\DataGridBundle\Grid\Source\Entity;
 use Ojs\Common\Helper\ActionHelper;
+use Symfony\Component\Security\Core\Exception\TokenNotFoundException;
 
 /**
  * Citation controller.
@@ -45,6 +46,8 @@ class CitationController extends Controller
         });
         $grid = $this->get('grid')->setSource($source);
         $actionColumn = new ActionsColumn("actions", 'actions');
+        ActionHelper::setup($this->get('security.csrf.token_manager'));
+
         $rowAction[] = ActionHelper::showAction('citation_show', 'id');
         $rowAction[] = ActionHelper::editAction('citation_edit', 'id');
         $rowAction[] = ActionHelper::deleteAction('citation_delete', 'id');
@@ -220,16 +223,22 @@ class CitationController extends Controller
     /**
      * Deletes a Citation entity.
      *
+     * @param $request
      * @param $id
      * @return RedirectResponse
      */
-    public function deleteAction($id)
+    public function deleteAction(Request $request, $id)
     {
         $em = $this->getDoctrine()->getManager();
         $entity = $em->getRepository('OjsJournalBundle:Citation')->find($id);
         if (!$entity) {
             throw $this->createNotFoundException('notFound');
         }
+
+        $csrf = $this->get('security.csrf.token_manager');
+        $token = $csrf->getToken('application_journal'.$id);
+        if($token!=$request->get('_token'))
+            throw new TokenNotFoundException("Token Not Found!");
         $em->remove($entity);
         $em->flush();
         $this->successFlashBag('successful.remove');
